Handle all incoming Inspector messages on main thread (#41503)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/41503

Changelog: [Internal]

Updates RCTInspectorPackagerConnection to dispatch all messages on the main thread (which is SocketRocket's default behaviour). This is in support of a new CDP implementation in React Native that will strictly assume it's called on the main thread (thus avoiding the need for explicit locking in many places).

Reviewed By: javache

Differential Revision: D51346659

fbshipit-source-id: c529b0aea97f7732cea58a4dc66993c5c8259958